hoisted google calendar fetch to route level
